requirements.txt
# Core Framework
langgraph>=0.0.50
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# Medical & Scientific
lifelines>=0.28.0
scipy>=1.12.0
scikit-learn>=1.4.0
xgbse>=0.4.0
shap>=0.44.0

# Knowledge Fabric & Graph
networkx>=3.2.0
rdflib>=7.0.0
jsonpickle>=3.0.0

# Data Processing & APIs
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
requests-cache>=1.2.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Parallel Processing
joblib>=1.3.0
multiprocess>=0.70.0
tqdm>=4.66.0

# Pathology & Imaging
tiffslide>=2.3.0
tiatoolbox>=1.4.0
scikit-image>=0.21.0
opencv-python>=4.9.0
Pillow>=10.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.20.0
streamlit>=1.30.0
streamlit-agraph>=0.0.45

# Notebook & Interactive
jupyter>=1.0.0
ipywidgets>=8.0.0
ipython>=8.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0
diskcache>=5.6.0
